+++ math/src/elliptic_curve/short_weierstrass/curves/bn_254/mod.rs
@@ -2,7 +2,4 @@ pub mod curve;
 pub mod default_types;
 pub mod field_extension;
 pub mod pairing;
-pub mod pairing_arkworks;
-pub mod pairing_post;
-pub mod pairing_zksync;
 pub mod twist;
